requirements.txt
google-generativeai>=0.4.0
pydantic>=2.0.0
tenacity>=8.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0

pyvis>=0.3.2
